Business rule applied

+++ gdct-app/backend/src/services/Program/service.js
@@ -1,10 +1,12 @@
 import Container from 'typedi';
 
 import ProgramRepository from '../../repositories/Program';
+import MasterValueRepository from '../../repositories/MasterValue'
 
 export default class ProgramService {
   constructor() {
     this.programRepository = Container.get(ProgramRepository);
+    this.masterValueRepository = Container.get(MasterValueRepository);
   }
 
   async createProgram(program) {
@@ -12,6 +14,11 @@ export default class ProgramService {
   }
 
   async deleteProgram(id) {
+
+    // Apply buisness rule, if program is referenced in mastervalue, throw an error to prevent deletion
+    const masterValueEntry = await this.masterValueRepository.findOneByProgramId(id);
+    if (masterValueEntry != null) throw new Error("This program is referenced in the master value table");
+
     return this.programRepository.delete(id);
   }
