+++ gdct-app/backend/src/services/TemplatePackage/service.js
@@ -1,10 +1,12 @@
 import Container from 'typedi';
 import TemplatePackageRepository from '../../repositories/TemplatePackage';
+import StatusRepository from '../../repositories/Status';
 
 // @Service()
 export default class TemplatePackageService {
   constructor() {
     this.templatePackageRepository = Container.get(TemplatePackageRepository);
+    this.statusRepository = Container.get(StatusRepository)
   }
 
   async createTemplatePackage(templatePackage) {
@@ -12,6 +14,8 @@ export default class TemplatePackageService {
   }
 
   async deleteTemplatePackage(id) {
+    const targetPackageStaus = await this.templatePackageRepository.findStatusById(id);
+    if (targetPackageStaus.name === "Published") throw new Error("This Template Package was already published");
     return this.templatePackageRepository.delete(id);
   }
